Relocate fundAccount operation
