+++ nodes/Stellar/actions/fundAccount/fundAccount/error/InvalidPublicKeyError.ts
